fix(postgres,redis,shared,testing): replay seq()/gap()/truncation reads cast COALESCE(seq,0)::bigint so a hot topic past 2^31 no longer raises 22003 server-side; abandoned redis groups self-expire via a 2x-memberTtl PEXPIRE margin on the members and meta hashes, stamped at join and refreshed by the heartbeat while the closed tombstone stays un-expired; the schema-drift guard verifies columns on the clean CREATE TABLE IF NOT EXISTS path and not only the 42P07 race, and scopes the information_schema lookup to current_schema(); table-name option capped at 42 bytes so generated idx_<table>_<suffix> identifiers stay distinct under Postgres's 63-byte limit; mock-pg records CREATE and ALTER columns and answers information_schema.columns so the drift guard verifies under unit tests; primary-key convention and its three natural-key exceptions documented

+++ src/postgres/replay.js
@@ -136,6 +136,10 @@ export function createReplay(client, options = {}) {
 		// Right-to-erasure column (ALTER so existing buffers forward-migrate).
 		await safeCreate(client, `ALTER TABLE ${table} ADD COLUMN IF NOT EXISTS user_id TEXT`);
 		await safeCreate(client, `CREATE INDEX IF NOT EXISTS idx_${table}_user ON ${table} (user_id)`);
+		// `topic` is the natural primary key: one counter row per topic, so a
+		// surrogate `_id` would add no identity. A documented divergence from the
+		// `<tablename>_id` convention the event table (svti_replay) follows - the
+		// same single-row-per-key shape as svti_alarms and svti_idempotency.
 		await safeCreate(client, `
 			CREATE TABLE IF NOT EXISTS ${seqTable} (
 				topic TEXT   PRIMARY KEY,
@@ -515,7 +519,7 @@ export function createReplay(client, options = {}) {
 				await ensureTable();
 				return client.query({
 					name: 'replay_seq_' + table,
-					text: `SELECT COALESCE(seq, 0)::int AS current_seq
+					text: `SELECT COALESCE(seq, 0)::bigint AS current_seq
 					         FROM ${seqTable}
 					        WHERE topic = $1`,
 					values: [topic]
@@ -553,7 +557,7 @@ export function createReplay(client, options = {}) {
 
 				const seqRes = await client.query({
 					name: 'replay_seq_' + table,
-					text: `SELECT COALESCE(seq, 0)::int AS current_seq
+					text: `SELECT COALESCE(seq, 0)::bigint AS current_seq
 					         FROM ${seqTable}
 					        WHERE topic = $1`,
 					values: [topic]
@@ -631,7 +635,7 @@ export function createReplay(client, options = {}) {
 				try {
 					const seqRes = await client.query({
 						name: 'replay_seq_' + table,
-						text: `SELECT COALESCE(seq, 0)::int AS current_seq
+						text: `SELECT COALESCE(seq, 0)::bigint AS current_seq
 						         FROM ${seqTable}
 						        WHERE topic = $1`,
 						values: [topic]
